// src/models/Product.ts
import { Dimensions } from "./Dimensions";
import { Metadata } from "./Metadata";
import { Review } from "./Review";


export interface Product {
  id: number;
  title: string;
  description: string;
  category: string;
  price: number;
  discountPercentage?: number;
  rating?: number;
  stock: number;
  tags: string[];
  brand: string;
  sku: string;
  weight: number;
  dimensions: Dimensions;
  warrantyInformation: string;
  shippingInformation: string;
  availabilityStatus: string;
  reviews: Review[];
  returnPolicy: string;
  minimumOrderQuantity: number;
  meta: Metadata;
  images: string[];
  thumbnail: string;
}